+++ Utilities/LegacyProjectElements.cs
@@ -79,7 +79,19 @@ public static class LegacyProjectElements
         "$(MSBuildBinPath)\\Microsoft.CSharp.targets",
         "$(MSBuildBinPath)\\Microsoft.VisualBasic.targets",
         "$(VSToolsPath)\\TeamTest\\Microsoft.TestTools.targets",
-        "$(MSBuildExtensionsPath)\\$(MSBuildToolsVersion)\\Microsoft.Common.props"
+        "$(MSBuildExtensionsPath)\\$(MSBuildToolsVersion)\\Microsoft.Common.props",
+        "$(VSToolsPath)\\WebApplications\\Microsoft.WebApplication.targets",
+        "$(SolutionDir)\\.nuget\\NuGet.targets",
+        "$(MSBuildExtensionsPath32)\\Microsoft\\VisualStudio\\v$(VisualStudioVersion)\\WebApplications\\Microsoft.WebApplication.targets",
+        "$(MSBuildExtensionsPath)\\Microsoft\\VisualStudio\\v$(VisualStudioVersion)\\WebApplications\\Microsoft.WebApplication.targets",
+        "$(MSBuildExtensionsPath32)\\Microsoft\\VisualStudio\\v10.0\\WebApplications\\Microsoft.WebApplication.targets",
+        "$(MSBuildExtensionsPath32)\\Microsoft\\VisualStudio\\v11.0\\WebApplications\\Microsoft.WebApplication.targets",
+        "$(MSBuildExtensionsPath32)\\Microsoft\\VisualStudio\\v12.0\\WebApplications\\Microsoft.WebApplication.targets",
+        "$(MSBuildExtensionsPath32)\\Microsoft\\VisualStudio\\v14.0\\WebApplications\\Microsoft.WebApplication.targets",
+        "$(MSBuildExtensionsPath32)\\Microsoft\\VisualStudio\\v15.0\\WebApplications\\Microsoft.WebApplication.targets",
+        "$(MSBuildExtensionsPath32)\\Microsoft\\VisualStudio\\v16.0\\WebApplications\\Microsoft.WebApplication.targets",
+        "$(VSToolsPath)\\TypeScript\\Microsoft.TypeScript.targets",
+        "$(MSBuildExtensionsPath32)\\Microsoft\\VisualStudio\\v$(VisualStudioVersion)\\TypeScript\\Microsoft.TypeScript.targets"
     };
 
     public static readonly HashSet<string> ProblematicTargets = new(StringComparer.OrdinalIgnoreCase)